Add plugins
Copying over code from petribrowser that emulates ssb plugin and
flumeview functionality so that you can load new views streaming over a
channel. That code was in turn ported directly from flumedb and adapted
to work with streams rather than a flumelog. Still need to implement
wrapping the view so that it waits to sync before calling back. Also
needs tests which will be next.